backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
Pillow==11.1.0
requests==2.32.3
deepface==0.0.93
opencv-python==4.11.0.86
face-recognition==1.3.0
mtcnn==1.0.0
pandas==2.2.3

